linting

+++ src/shader_variables.rs
@@ -17,8 +17,8 @@ impl ShaderVariables {
             screen_width: SCREEN_WIDTH as f32,
             screen_height: SCREEN_HEIGHT as f32,
             mode: 0.0,
-            scanline_interval: 6.0 as f32,
-            mask_size: 3.0 as f32,
+            scanline_interval: 6.0,
+            mask_size: 3.0,
             mask_type: 3.0,
             horiz_distortion: 0.0,
             vert_distortion: 0.0
